pyproject.toml
[tool.poetry]
name = "idc-annotation-conversion"
version = "0.1.0"
description = "Python code for converting TCGA annotations into DICOM format."
authors = ["Chris Bridge <cbridge@mgh.harvard.edu>"]
readme = "README.md"
packages = [{include = "idc_annotation_conversion", from = "src" }]

[tool.poetry.dependencies]
python = "^3.9"
pydicom = "^2.3.1"
highdicom = "^0.21.0"
google-cloud-bigquery = "^3.6.0"
google-cloud-storage = "^2.7.0"
google-auth = "^2.16.2"
pandas = "^1.5.3"
db-dtypes = "^1.0.5"
mahotas = "^1.4.13"
shapely = "^2.0.1"
click = "^8.1.3"
dicomweb-client = "^0.59.1"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
mypy = "^1.1.1"
ipython = "^8.11.0"

[tool.poetry.scripts]
idc-annotation-conversion = "idc_annotation_conversion.__main__:run"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
